+++ testdata/bless_provider_yaml/terraform/accounts/foo/fogg.tf
@@ -80,14 +80,7 @@ terraform {
     random = {
       source = "hashicorp/random"
 
-      version = "~> 2.2"
-
-    }
-
-    template = {
-      source = "hashicorp/template"
-
-      version = "~> 2.2"
+      version = "~> 3.4"
 
     }